+++ internal/api/api.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/asashakira/mai.gg-api/internal/api/handler"
+	"github.com/asashakira/mai.gg/internal/api/handler"
 	"github.com/go-chi/chi/v5"
 )
