+++ internal/api/handler/beatmaps.go
@@ -7,7 +7,7 @@ import (
 	"log"
 	"net/http"
 
-	database "github.com/asashakira/mai.gg-api/internal/database/sqlc"
+	database "github.com/asashakira/mai.gg/internal/database/sqlc"
 	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
