+++ internal/api/handler/songs.go
@@ -9,7 +9,7 @@ import (
 	"net/url"
 	"time"
 
-	database "github.com/asashakira/mai.gg-api/internal/database/sqlc"
+	database "github.com/asashakira/mai.gg/internal/database/sqlc"
 	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
