+++ internal/api/handler/user_scrape_metadata.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"time"
 
-	database "github.com/asashakira/mai.gg-api/internal/database/sqlc"
+	database "github.com/asashakira/mai.gg/internal/database/sqlc"
 	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5/pgtype"
