+++ internal/api/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 import (
-	database "github.com/asashakira/mai.gg-api/internal/database/sqlc"
+	database "github.com/asashakira/mai.gg/internal/database/sqlc"
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5/pgtype"
 )
